# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estuaire Emploie API Documentation",
        "description": "Documentation de l'API REST pour l'application mobile Estuaire Emploie - Plateforme de recherche d'emploi au Cameroun",
        "contact": {
            "email": "contact@estuaire-emploie.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API Principal"
        }
    ],
    "paths": {
...
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Liste des offres favorites",
                "description": "Récupère la liste paginée de toutes les offres d'emploi mises en favori par l'utilisateur connecté",
                "operationId": "getFavorites",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des favoris récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Développeur Full Stack"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "salary_min": {
                                                        "type": "string",
                                                        "example": "500000"
                                                    },
                                                    "salary_max": {
                                                        "type": "string",
                                                        "example": "800000"
                                                    },
                                                    "salary_negotiable": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "experience_level": {
                                                        "type": "string",
                                                        "example": "intermediaire"
                                                    },
                                                    "is_featured": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "views_count": {
                                                        "type": "integer",
                                                        "example": 150
                                                    },
                                                    "published_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "company": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "logo": {
                                                                "type": "string"
                                                            },
                                                            "is_verified": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "category": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "location": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "contract_type": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 45
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/jobs/{job}/favorite": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Ajouter/Retirer une offre des favoris",
                "description": "Toggle le statut favori d'une offre d'emploi. Si l'offre est déjà en favori, elle sera retirée. Sinon, elle sera ajoutée.",
                "operationId": "toggleFavorite",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut favori mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ajouté aux favoris"
                                        },
                                        "is_favorite": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Offre d'emploi non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/jobs/{job}/is-favorite": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Vérifier si une offre est en favori",
                "description": "Vérifie si une offre d'emploi spécifique est dans les favoris de l'utilisateur connecté",
                "operationId": "isFavorite",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut favori de l'offre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_favorite": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Offre d'emploi non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum - Utiliser le token obtenu lors du login/register",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}